# src/ports/api/account/dtos/account.dto.ts
export class AccountDto {
  id: string;
  cashAvailableForWithdrawal: number;
  cashBalance: number;
  created_at: Date;
  updated_at: Date;

  constructor(partial: Partial<AccountDto>) {
    Object.assign(this, partial);
  }
}